Fix [tree] y dtype=intp

+++ tests/set_up_variables_and_imports.py
@@ -28,4 +28,4 @@ iris.target = iris.target[perm]
 
 X = iris.data
 y = iris.target
-y = np.ascontiguousarray(y, dtype=np.float64)
+y = np.ascontiguousarray(y, dtype=np.intp)
